+++ neoforge/src/main/java/com/fej1fun/potentials/neoforge/capabilities/holders/EnergyBlockHolder.java
@@ -61,11 +61,11 @@ public class EnergyBlockHolder implements NoProviderBlockCapabilityHolder<Univer
         registeredBlocks.forEach(block ->
                 event.registerBlock(Capabilities.Energy.BLOCK, (level, pos, state, blockEntity, direction) -> {
                     if (blockEntity instanceof EnergyProvider.BLOCK energyBlock) {
-                            UniversalEnergyStorage energy = energyBlock.getEnergy(direction);
-                            return energy == null ? null : new NeoForgeEnergyStorage(energy);
+                        UniversalIEnergyStorage energy = (UniversalIEnergyStorage) energyBlock.getEnergy(direction);
+                        return energy == null ? null : new NeoForgeEnergyStorage(energy);
                     }
                     if (state.getBlock() instanceof EnergyProvider.BLOCK energyBlock) {
-                        UniversalEnergyStorage energy = energyBlock.getEnergy(direction);
+                        UniversalIEnergyStorage energy = (UniversalIEnergyStorage) energyBlock.getEnergy(direction);
                         return energy == null ? null : new NeoForgeEnergyStorage(energy);
                     }
                     return null;
@@ -73,7 +73,7 @@ public class EnergyBlockHolder implements NoProviderBlockCapabilityHolder<Univer
         registeredBlockEntities.forEach(type ->
                 event.registerBlockEntity(Capabilities.Energy.BLOCK, type.get(), ((blockEntity, direction) -> {
                     if (blockEntity instanceof EnergyProvider.BLOCK energyBlock) {
-                        UniversalEnergyStorage energy = energyBlock.getEnergy(direction);
+                        UniversalIEnergyStorage energy = (UniversalIEnergyStorage) energyBlock.getEnergy(direction);
                         return energy == null ? null : new NeoForgeEnergyStorage(energy);
                     }
                     return null;