+++ neoforge/src/main/java/com/fej1fun/potentials/neoforge/energy/NeoForgeEnergyStorage.java
@@ -1,15 +1,13 @@
 package com.fej1fun.potentials.neoforge.energy;
 
-import com.fej1fun.potentials.energy.UniversalEnergyStorage;
 import net.neoforged.neoforge.transfer.energy.EnergyHandler;
-import net.neoforged.neoforge.transfer.transaction.Transaction;
 import net.neoforged.neoforge.transfer.transaction.TransactionContext;
 import org.jetbrains.annotations.NotNull;
 
 public class NeoForgeEnergyStorage implements EnergyHandler {
-    final UniversalEnergyStorage universalEnergyStorage;
+    final UniversalIEnergyStorage universalEnergyStorage;
 
-    public NeoForgeEnergyStorage(@NotNull UniversalEnergyStorage universalEnergyStorage) {
+    public NeoForgeEnergyStorage(@NotNull UniversalIEnergyStorage universalEnergyStorage) {
        this.universalEnergyStorage = universalEnergyStorage;
     }
 
@@ -25,11 +23,11 @@ public class NeoForgeEnergyStorage implements EnergyHandler {
 
     @Override
     public int insert(int toReceive, TransactionContext transactionContext) {
-        return universalEnergyStorage.insert(toReceive, false);
+        return universalEnergyStorage.insert(toReceive, false, transactionContext);
     }
 
     @Override
     public int extract(int toReceive, TransactionContext transactionContext) {
-       return  universalEnergyStorage.extract(toReceive, false);
+       return  universalEnergyStorage.extract(toReceive, false, transactionContext);
     }
 }