+++ changelog.md
@@ -3,3 +3,4 @@
 
 ## Fix
 - Fix setFluidInTank doesn't update the component
+- Pass transaction context on neoforge during insert/extract